pkg/idtools: remove user/group lookup functions
These functions were written before Go had support for native user and
group lookups. Nowadays there is a cgo implementation in package os/user
which will yield the same answers as the getent command as it calls the
same libc lookup functions. The os/user package will fall back to a
pure-Go implementation which parses /etc/passwd and /etc/group when the
binary is built with cgo disabled so there is no need to rely on
libcontainer's parsers.

As a consequence of removing the getent command fallback, static builds
will no longer be able to resolve user or group names defined in
databases outside the filesystem.

Signed-off-by: Cory Snider <csnider@mirantis.com>