Fix secret mounting with userns_remap enabled
The previous commit here bfedd2725971303efb7a2fe5d6990317b381622f modified
the owner and permission of the DOCKER_TMP to `root:root` and `700`
respectively, and it seems to have assumed that the folder is not
needed by the remapped root user, however, this is not actually the case.
E.g. container secret mounting function creates temp secret files inside
this folder.

Signed-off-by: Jiashu Chen <cjs20080808@hotmail.com>