Add support for nftables
With the release of EL8, nftables has become the default firewall
implementation on minimal installs of CentOS/OEL/RHEL8. While an
iptables driver is available (both in libnetwork and a mapper in
nftables itself), this is not guaranteed to be present. firewalld
pulls in iptables, but firewalld itself does not yet support raw
commands to nftables.

iptables calls were heavily enmeshed in libnetwork. It was
necessary to create a relatively large interface in order to
break the coupling to native iptables implementations, but this
offers the flexibility to support or add other drivers in the
future

Signed-off-by: Ryan Barry <rbarry@mirantis.com>