rootless: remove redundant sync.Once
This was added in ec87479b7e2bf6f1b5bcc657a377c6e6a847574f, but it's unclear
why a sync.Once was used just for reading an environment-variable. The
related PR had a lot of review comments, so perhaps an earlier implementation
used something more heavy-weight, or it was just overlooked.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>