daemon/archive: attempt to resolve leaking mounts during cp
This is a naive first attempt at resolving the issue reported in #38995
and #43390. This approach does pass the integration tests locally, and
it passes my minimal reproduction. It additionally passes the short
stress-test of the daemon.

However, the attached integration test, while correct, does not properly
fail, without this patch. This is because it is not triggered by
Docker-in-Docker -- to see this bug, we have to be in the 'true root'
mount namespace.

There are a couple other possibilities we can explore -- the first one I
see is an [alternate test location]. Testing the daemon without the
container abstraction would take this from a regression test for `docker
cp` to a test of the soundness of mount propagation during the special
case of mounting the daemon root into the container.

The second alternative is a completely different, larger-scale change,
to use a disposable mount namespace in the engine to do all container
mounting in daemon/archive. This is more invasive and harder to
prototype, but less likely to have issues in the future as we can
completely sidestep this category of issues (interactions with the root
mount namespace).

Eyeballs and input on this approach, or the merits of the other two
approaches (as well as anything I haven't forseen) is appreciated. I
would like feedback from upstream (especially on how to test/validate
this) before I go to far ahead with more revisions.

A prior attempt at solving this can be found at #38993.

  [alternate test location]: https://github.com/moby/moby/blob/c55a4ac7795c7606b548b38e24673733481e2167/daemon/daemon_linux_test.go

Signed-off-by: Bjorn Neergaard <bneergaard@mirantis.com>