volume/local: remove hack for downgrading docker 1.7 to 1.6
This was added in bd9814f0db9c8a087e42b66eabb8413bf1b2ab66 to support downgrading
docker 1.7 to 1.6.

The related migration code was removed in 0023abbad34282762d5bd17302776d2a8521fffc
(Docker 18.05), which was also the last consumer of VolumeDataPathName outside
of the package, so that const can be un-exported.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>